fix: deadcode drift after factory-graph-editor public barrel removal
Remove test-only exports, retarget factory-graph-editor Vitest imports from the deleted public barrel to owning hook/lib modules, and inline collectPendingRemovalEdgeIds in removals tests.
