+++ ui/src/features/factory-graph-editor/lib/factory-graph-operations.test.ts
@@ -1,17 +1,16 @@
 import { describe, expect, it } from "vitest";
 
+import { createEmptyFactoryGraphDraft } from "./factory-graph-draft-types";
+import { updateFactoryGraphNodeField } from "./factory-graph-field-operations";
 import {
   addFactoryGraphNode,
   applyFactoryGraphPendingEdits,
-  buildFactoryGraphSaveInput,
   buildFactoryGraphState,
   connectFactoryGraphNodes,
-  createEmptyFactoryGraphDraft,
   disconnectFactoryGraphEdge,
-  projectFactoryGraphToReactFlow,
   removeFactoryGraphNode,
-  updateFactoryGraphNodeField,
-} from "../public";
+} from "./factory-graph-operations";
+import { projectFactoryGraphToReactFlow } from "./factory-graph-react-flow-projection";
 import { baseFactoryDefinition } from "./factory-graph-draft.test-helpers";
 
 // biome-ignore lint/complexity/noExcessiveLinesPerFunction: keeps the operation contract scenarios co-located.
@@ -385,7 +384,7 @@ describe("factory graph operations", () => {
       baseFactoryDefinition,
       draft,
     });
-    const saveInput = buildFactoryGraphSaveInput({
+    const saveInput = applyFactoryGraphPendingEdits({
       baseFactoryDefinition,
       draft,
     });
@@ -404,7 +403,7 @@ function expectOk<T>(
   result: ReturnType<
     | typeof addFactoryGraphNode
     | typeof applyFactoryGraphPendingEdits
-    | typeof buildFactoryGraphSaveInput
+    | typeof applyFactoryGraphPendingEdits
     | typeof connectFactoryGraphNodes
     | typeof disconnectFactoryGraphEdge
     | typeof removeFactoryGraphNode
